chore: refactor auth utils
