+++ packages/utils/src/auth-utils.ts
@@ -1,25 +1,13 @@
-import _ from 'lodash';
-
-import {
-  API_ERROR,
-  ROLES,
-  TIME_EXPIRATION_7D,
-  DEFAULT_MIN_LIMIT_PASSWORD,
-} from '@verdaccio/dev-commons';
+import { ROLES, TIME_EXPIRATION_7D, DEFAULT_MIN_LIMIT_PASSWORD } from '@verdaccio/dev-commons';
 import {
   RemoteUser,
   AllowAccess,
   PackageAccess,
-  Callback,
-  Config,
   Security,
   APITokenOptions,
   JWTOptions,
-  IPluginAuth,
 } from '@verdaccio/types';
 import { VerdaccioError } from '@verdaccio/commons-api';
-
-import { ErrorCode } from './utils';
 
 export interface CookieSessionToken {
   expires: Date;
@@ -85,93 +73,16 @@ export type AllowActionCallback = (
   error: VerdaccioError | null,
   allowed?: AllowActionCallbackResponse
 ) => void;
+
 export type AllowAction = (
   user: RemoteUser,
   pkg: AuthPackageAllow,
   callback: AllowActionCallback
 ) => void;
+
 export interface AuthPackageAllow extends PackageAccess, AllowAccess {
   // TODO: this should be on @verdaccio/types
   unpublish: boolean | string[];
-}
-
-export type ActionsAllowed = 'publish' | 'unpublish' | 'access';
-
-export function allow_action(action: ActionsAllowed, logger): AllowAction {
-  return function allowActionCallback(
-    user: RemoteUser,
-    pkg: AuthPackageAllow,
-    callback: AllowActionCallback
-  ): void {
-    logger.trace({ remote: user.name }, `[auth/allow_action]: user: @{user.name}`);
-    const { name, groups } = user;
-    const groupAccess = pkg[action] as string[];
-    const hasPermission = groupAccess.some((group) => name === group || groups.includes(group));
-    logger.trace(
-      { pkgName: pkg.name, hasPermission, remote: user.name, groupAccess },
-      `[auth/allow_action]: hasPermission? @{hasPermission} for user: @{user}`
-    );
-
-    if (hasPermission) {
-      logger.trace({ remote: user.name }, `auth/allow_action: access granted to: @{user}`);
-      return callback(null, true);
-    }
-
-    if (name) {
-      callback(
-        ErrorCode.getForbidden(`user ${name} is not allowed to ${action} package ${pkg.name}`)
-      );
-    } else {
-      callback(
-        ErrorCode.getUnauthorized(`authorization required to ${action} package ${pkg.name}`)
-      );
-    }
-  };
-}
-
-/**
- *
- */
-export function handleSpecialUnpublish(logger): any {
-  return function (user: RemoteUser, pkg: AuthPackageAllow, callback: AllowActionCallback): void {
-    const action = 'unpublish';
-    // verify whether the unpublish prop has been defined
-    const isUnpublishMissing: boolean = _.isNil(pkg[action]);
-    const hasGroups: boolean = isUnpublishMissing ? false : (pkg[action] as string[]).length > 0;
-    logger.trace(
-      { user: user.name, name: pkg.name, hasGroups },
-      `fallback unpublish for @{name} has groups: @{hasGroups} for @{user}`
-    );
-
-    if (isUnpublishMissing || hasGroups === false) {
-      return callback(null, undefined);
-    }
-
-    logger.trace(
-      { user: user.name, name: pkg.name, action, hasGroups },
-      `allow_action for @{action} for @{name} has groups: @{hasGroups} for @{user}`
-    );
-    return allow_action(action, logger)(user, pkg, callback);
-  };
-}
-
-export function getDefaultPlugins(logger: any): IPluginAuth<Config> {
-  return {
-    authenticate(user: string, password: string, cb: Callback): void {
-      cb(ErrorCode.getForbidden(API_ERROR.BAD_USERNAME_PASSWORD));
-    },
-
-    adduser(user: string, password: string, cb: Callback): void {
-      return cb(ErrorCode.getConflict(API_ERROR.BAD_USERNAME_PASSWORD));
-    },
-
-    // FIXME: allow_action and allow_publish should be in the @verdaccio/types
-    // @ts-ignore
-    allow_access: allow_action('access', logger),
-    // @ts-ignore
-    allow_publish: allow_action('publish', logger),
-    allow_unpublish: handleSpecialUnpublish(logger),
-  };
 }
 
 export function createSessionToken(): CookieSessionToken {
