+++ packages/utils/test/auth-utils.spec.ts
@@ -1,16 +1,13 @@
 import { API_ERROR, ROLES } from '@verdaccio/dev-commons';
 import { VerdaccioError, getForbidden } from '@verdaccio/commons-api';
 import {
-  allow_action,
   createAnonymousRemoteUser,
   createRemoteUser,
   validatePassword,
-  ActionsAllowed,
-  AllowActionCallbackResponse,
-  getDefaultPlugins,
   createSessionToken,
   getAuthenticatedMessage,
 } from '../src';
+
 jest.mock('@verdaccio/logger', () => ({
   logger: { trace: jest.fn() },
 }));
@@ -60,123 +57,10 @@ describe('Auth Utilities', () => {
     });
   });
 
-  describe('allow_action', () => {
-    describe('access/publish/unpublish and anonymous', () => {
-      const packageAccess = {
-        name: 'foo',
-        version: undefined,
-        access: ['foo'],
-        unpublish: false,
-      };
-
-      // const type = 'access';
-      test.each(['access', 'publish', 'unpublish'])(
-        'should restrict %s to anonymous users',
-        (type) => {
-          allow_action(type as ActionsAllowed, { trace: jest.fn() })(
-            createAnonymousRemoteUser(),
-            {
-              ...packageAccess,
-              [type]: ['foo'],
-            },
-            (error: VerdaccioError | null, allowed: AllowActionCallbackResponse) => {
-              expect(error).not.toBeNull();
-              expect(allowed).toBeUndefined();
-            }
-          );
-        }
-      );
-
-      test.each(['access', 'publish', 'unpublish'])(
-        'should allow %s to anonymous users',
-        (type) => {
-          allow_action(type as ActionsAllowed, { trace: jest.fn() })(
-            createAnonymousRemoteUser(),
-            {
-              ...packageAccess,
-              [type]: [ROLES.$ANONYMOUS],
-            },
-            (error: VerdaccioError | null, allowed: AllowActionCallbackResponse) => {
-              expect(error).toBeNull();
-              expect(allowed).toBe(true);
-            }
-          );
-        }
-      );
-
-      test.each(['access', 'publish', 'unpublish'])(
-        'should allow %s only if user is anonymous if the logged user has groups',
-        (type) => {
-          allow_action(type as ActionsAllowed, { trace: jest.fn() })(
-            createRemoteUser('juan', ['maintainer', 'admin']),
-            {
-              ...packageAccess,
-              [type]: [ROLES.$ANONYMOUS],
-            },
-            (error: VerdaccioError | null, allowed: AllowActionCallbackResponse) => {
-              expect(error).not.toBeNull();
-              expect(allowed).toBeUndefined();
-            }
-          );
-        }
-      );
-
-      test.each(['access', 'publish', 'unpublish'])(
-        'should allow %s only if user is anonymous match any other groups',
-        (type) => {
-          allow_action(type as ActionsAllowed, { trace: jest.fn() })(
-            createRemoteUser('juan', ['maintainer', 'admin']),
-            {
-              ...packageAccess,
-              [type]: ['admin', 'some-other-group', ROLES.$ANONYMOUS],
-            },
-            (error: VerdaccioError | null, allowed: AllowActionCallbackResponse) => {
-              expect(error).toBeNull();
-              expect(allowed).toBe(true);
-            }
-          );
-        }
-      );
-
-      test.each(['access', 'publish', 'unpublish'])(
-        'should not allow %s anonymous if other groups are defined and does not match',
-        (type) => {
-          allow_action(type as ActionsAllowed, { trace: jest.fn() })(
-            createRemoteUser('juan', ['maintainer', 'admin']),
-            {
-              ...packageAccess,
-              [type]: ['bla-bla-group', 'some-other-group', ROLES.$ANONYMOUS],
-            },
-            (error: VerdaccioError | null, allowed: AllowActionCallbackResponse) => {
-              expect(error).not.toBeNull();
-              expect(allowed).toBeUndefined();
-            }
-          );
-        }
-      );
-    });
-  });
   describe('createSessionToken', () => {
     test('should generate session token', () => {
       expect(createSessionToken()).toHaveProperty('expires');
       expect(createSessionToken().expires).toBeInstanceOf(Date);
-    });
-  });
-
-  describe('getDefaultPlugins', () => {
-    test('authentication should fail by default (default)', () => {
-      const plugin = getDefaultPlugins({ trace: jest.fn() });
-      plugin.authenticate('foo', 'bar', (error: any) => {
-        expect(error).toEqual(getForbidden(API_ERROR.BAD_USERNAME_PASSWORD));
-      });
-    });
-
-    test('add user should fail by default (default)', () => {
-      const plugin = getDefaultPlugins({ trace: jest.fn() });
-      // @ts-ignore
-      plugin.adduser('foo', 'bar', (error: any) => {
-        expect(error).toEqual(getForbidden(API_ERROR.BAD_USERNAME_PASSWORD));
-      });
     });
   });
